Improve squashing alogrithm to account for stair stepping topography/bathymetry.
Improve squashing algorithm to account for topography/bathymetry cutting
through the bottom half of octants. These octants are not included in
the Etree, so queries just below the topography/bathymetry surface would
not return data. The remedy is to detect these cases and return the
values in the octant directly below the desired one.

Ultimately, converting the rasterized model from uniform values within a
cell to interpolated values would avert this issue.